pyproject.toml
[tool.poetry]
name = "codepic"
version = "0.1.0"
description = "code picture"
authors = ["Thomas Harrison <theharrisoncrafter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygments = "^2.19.1"
pillow = "^11.1.0"
click = "^8.1.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
ruff = "^0.8.4"
ipykernel = "^6.29.5"

[tool.poetry.scripts]
codepic = "codepic:cli"

[tool.ruff.lint]
extend-select = ["I"]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--cov=src --cov-branch --cov-fail-under=100"

[tool.coverage.run]
branch = false
source = ["src"]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
